Add separate submodule for Gst utils
